remove compile flags

+++ src/nnue.cpp
@@ -31,7 +31,7 @@ static void decode_bullet_affine(const uint8_t* p, T* out) {
     const T* flat = reinterpret_cast<const T*>(p);
     for (int i = 0; i < NUM_IN; ++i)
         for (int o = 0; o < NUM_OUT; ++o)
-            out[o * NUM_IN + i] = flat[(size_t)i * NUM_OUT + o];
+            out[o * NUM_IN + i] = flat[(size_t)i * NUM_OUT + (size_t)o];
 }
 
 // ============================================================
@@ -396,7 +396,7 @@ int NNUE::eval_screlu(bool is_white_move, U64 occ) {
     sum += (int64_t)l1b_small[bucket];
     sum *= SCALE;
     sum /= int64_t(qa_small * QB);
-    return sum;
+    return (int)sum;
 }
 
 int NNUE::eval_screlu_simd(bool is_white_move, U64 occ) {
